+++ game/campaignloader/__init__.py
@@ -1,1 +1,2 @@
 from .campaign import Campaign
+from .campaignairwingconfig import CampaignAirWingConfig, SquadronConfig
