+++ game/squadrons/__init__.py
@@ -1,0 +1,3 @@
+from .airwing import AirWing
+from .pilot import Pilot
+from .squadron import Squadron
