--- gen/__init__.py
@@ -1,13 +1,0 @@
-from .aircraft import *
-from .armor import *
-from .airsupportgen import *
-from .conflictgen import *
-from .visualgen import *
-from .triggergen import *
-from .environmentgen import *
-from .groundobjectsgen import *
-from .briefinggen import *
-from .forcedoptionsgen import *
-from .kneeboard import *
-
-from . import naming
